+++ apps/cli/vite.config.ts
@@ -1,6 +1,7 @@
 import { defineConfig, type UserConfig } from "vite";
 
-import pubConfig from "../../vite.config.pub";
+import pubConfig, { root } from "../../vite.config.pub";
+import * as path from 'path';
 import { merge } from "lodash-es";
 
 export default defineConfig(merge<UserConfig, UserConfig>(
@@ -8,6 +9,9 @@ export default defineConfig(merge<UserConfig, UserConfig>(
   {
     server: {
       port: 1422
+    },
+    build: {
+      outDir: path.resolve(root, 'dist-cli'),
     }
   }
 ));