+++ apps/main/src/vite-env.d.ts
@@ -8,6 +8,7 @@ declare module "*.vue" {
 
 declare const __DEV__: boolean;
 declare const __WINDOWS__: boolean;
+declare const __LINUX__: boolean;
 declare const __DESKTOP__: string;
 declare const __LIGHT_MODE__: boolean;
 declare const __LOCAL_MODE__: boolean;